Rename files in gammapy/data to be more uniform
