+++ gammapy/data/__init__.py
@@ -6,16 +6,16 @@
 class InvalidDataError(Exception):
     """Invalid data found."""
 
+from .utils import *
 from .target import *
+from .pointing import *
 from .data_manager import *
 from .data_store import *
 from .event_list import *
 from .gti import *
 from .hdu_index_table import *
-from .observation import *
-from .obsgroup import *
 from .observers import *
-from .utils import *
-from .observation_summary import *
-from .pointing import *
-from .observation_stats import *
+from .obs_table import *
+from .obs_group import *
+from .obs_summary import *
+from .obs_stats import *
