+++ gammapy/data/obs_group.py
