+++ gammapy/data/obs_stats.py
