+++ gammapy/data/obs_table.py
