+++ gammapy/data/tests/test_obs_stats.py
