+++ gammapy/data/tests/test_obs_summary.py
