+++ gammapy/data/tests/test_obs_table.py
